login + register + tutor list + tutor detail

+++ lib/pages/tutors/widgets/star_rating.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class StarRating extends StatefulWidget {
-  StarRating({this.rating = 5, this.size = 14});
-  int? rating;
+  StarRating({this.rating = 5.0, this.size = 14.0});
+  double? rating; // Change from int to double
   double? size;
   @override
   _StarRatingState createState() => _StarRatingState();
@@ -18,7 +18,7 @@ class _StarRatingState extends State<StarRating> {
         return GestureDetector(
           onTap: () {
             setState(() {
-              widget.rating = index + 1;
+              widget.rating = index + 1.0; // Change to double
             });
           },
           child: Icon(
